requirements.txt
fastapi
uvicorn
sqlmodel
requests
yfinance
apscheduler
pymysql
python-dotenv
